domain primitive

+++ src/models.rs
@@ -6,8 +6,8 @@ pub struct Student {
     pub id: Option<i64>,
     pub stu_no: Option<String>,
     pub name: Option<String>,
-    pub age: Option<i8>,
-    pub class_id: Option<i64>,
+    pub age: Option<u8>,
+    pub class_id: Option<u32>,
     pub address: Option<String>,
 }
